requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
